Rename "functional tests" to "integration"
